add some missing refs, reorganize files
