+++ docs/source/user_guide/index.md
@@ -56,5 +56,4 @@ wrappers/index
 
 additional_resources/performance_benchmarking
 additional_resources/citation
-reference/index.rst
 ```
